support OpenTelemetry Http exporter and fix module data problem

+++ src/rpc_global.h
@@ -41,12 +41,12 @@ public:
 				   struct sockaddr_storage *ss, socklen_t *ss_len) const;
 
 	long long get_trace_id();
-	long long get_span_id() { return this->span_id++; }
+	unsigned int get_span_id() { return this->span_id++; }
 
 private:
 	SRPCGlobal();
 	SnowFlake snowflake;
-	std::atomic<long long> span_id;
+	std::atomic<unsigned int> span_id;
 };
 
 } // namespace srpc
